Moved styled-heading to styled-components folder
