+++ src/components/navigation/nav-header.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Navbar from './navbar';
-import StyledHeading from './styled-heading';
+import StyledHeading from '../../styled-components/styled-heading';
 
 export default function HeaderWithNav({ isAdmin, children }) {
   return (
